Move register, segment register, segment struct and related functions into kvm/registers.go
